Rewrite Song Selector UI Logic
Move into using a pooled object setup. Reduces the amount of required song buttons down to just want is need to be displayed at a given time, instead of using a new song button for every single level that is displayed.

This change makes the length of genres or categories have zero impact on performance.